requirements.txt
certifi==2020.6.20
Flask==1.1.2
Flask-Cors==3.0.9
ghostscript==0.6
googleapis-common-protos==1.52.0
gunicorn==20.0.4
pyquery==1.4.1
Pyrebase4==4.4.1
utils==1.0.1
xlrd==1.2.0
